+++ frontend/src/app/shared/models/confirmWindow/IConfirmDialogData.ts
@@ -1,4 +1,5 @@
-import { Type } from '@angular/core';
+import { INewMeetingMember } from '@core/models/INewMeetingTeamMember';
+import { LocationType } from '@shared/enums/locationType';
 import { IConfirmButtonOptions } from '@shared/models/confirmWindow/IConfirmButtonOptions';
 
 export interface IConfirmDialogData {
@@ -6,5 +7,10 @@ export interface IConfirmDialogData {
     message?: string;
     buttonsOptions: IConfirmButtonOptions[];
     titleImagePath?: string;
-    component?: Type<any>;
+    dateTime?: Date;
+    duration?: number;
+    meetingName?: string;
+    participants?: INewMeetingMember[];
+    location?: LocationType;
+    link?: string;
 }
