+++ frontend/src/app/shared/shared.module.ts
@@ -8,7 +8,7 @@ import { DomSanitizer } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 import { NgxMaskModule } from 'ngx-mask';
 
-import { BookingWindowComponent } from './components/confirmation-window/booking-window/booking-window.component';
+import { BookingWindowComponent } from './components/booking-window/booking-window.component';
 import { ConfirmationWindowComponent } from './components/confirmation-window/confirmation-window.component';
 import { LoadingSpinnerComponent } from './components/loading-spinner/loading-spinner.component';
 import { NotFoundComponent } from './components/not-found/not-found.component';
